+++ src/RutokenPkcs11InteropTests/HighLevelAPI/_HL_29_ManageFlashTest.cs
@@ -61,10 +61,10 @@ namespace RutokenPkcs11InteropTests.HighLevelAPI
 
                 // Полное удаление информации с флеш-памяти с последующим созданием
                 // разделов в соответствии с переданными параметрами
-                uint volumeRwSize = driveSize / 2;
-                uint volumeRoSize = driveSize / 4;
-                uint volumeHiSize = driveSize / 8;
-                uint volumeCdSize = driveSize - volumeRwSize - volumeRoSize - volumeHiSize;
+                ulong volumeRwSize = driveSize / 2;
+                ulong volumeRoSize = driveSize / 4;
+                ulong volumeHiSize = driveSize / 8;
+                ulong volumeCdSize = driveSize - volumeRwSize - volumeRoSize - volumeHiSize;
 
                 var initParams = new List<VolumeFormatInfoExtended>()
                 {
